+++ SwiftBook/ViewControllers/EmojihubViewController.swift
@@ -6,6 +6,7 @@
 //
 
 import UIKit
+
 import SpringAnimation
 
 final class EmojihubViewController: UIViewController {
@@ -51,7 +52,7 @@ final class EmojihubViewController: UIViewController {
             case .success(let emojihub):
                 self?.activityIndicator.stopAnimating()
                 self?.emojihub = emojihub
-                self?.animate()
+                AnimationManager.share.animate(sender: self?.emojiLabel ?? SpringLabel())
             case .failure(let error):
                 print(error.localizedDescription)
                 AlertManager.shared.showAlert(from: self, status: .failed)
@@ -61,52 +62,4 @@ final class EmojihubViewController: UIViewController {
     
 }
 
-// MARK: Smile Animation
 
-extension EmojihubViewController {
-    
-    func animate() {
-        
-        let nextAnimation = randomPreset()
-        
-        emojiLabel.animation = nextAnimation.animation
-        emojiLabel.curve = nextAnimation.curve
-        emojiLabel.force = nextAnimation.force
-        emojiLabel.duration = nextAnimation.duration
-        emojiLabel.delay = nextAnimation.delay
-        emojiLabel.animate()
-        
-    }
-    
-    func randomPreset() -> (animation: String, curve: String, force: CGFloat, duration: CGFloat, delay: CGFloat) {
-        var animations = AnimationPreset.allCases
-        
-        if let fadeOutIndex = AnimationPreset.allCases.firstIndex(of: AnimationPreset.fadeOut) {
-            animations.remove(at: fadeOutIndex)
-        }
-        if let zoomOutIndex = AnimationPreset.allCases.firstIndex(of: AnimationPreset.zoomOut) {
-            animations.remove(at: zoomOutIndex)
-        }
-        
-        if let fallIndex = AnimationPreset.allCases.firstIndex(of: AnimationPreset.fall) {
-            animations.remove(at: fallIndex)
-        }
-        
-        let curves = AnimationCurve.allCases
-        
-        var force: CGFloat {
-            CGFloat.random(in: 1...2)
-        }
-        
-        var duration: CGFloat {
-            CGFloat.random(in: 0.5...2)
-        }
-        
-        var delay: CGFloat {
-            CGFloat.random(in: 0...1)
-        }
-        
-        return (animation: animations.randomElement()?.rawValue ?? "", curve: curves.randomElement()?.rawValue ?? "", force: force, duration: duration, delay: delay)
-    }
-    
-}
